more autodiff functions

+++ test/test_autodiff/test_nonlinear/test_logit.py
@@ -8,7 +8,7 @@ from prml import autodiff
 class TestLogit(unittest.TestCase):
 
     def test_logit(self):
-        npx = np.random.uniform(0, 1, (3, 5))
+        npx = np.random.uniform(0, 1, (3, 5)).astype(np.float32)
         x = autodiff.asarray(npx)
         y = autodiff.logit(x)
         self.assertTrue(np.allclose(y.value, np.log(npx / (1 - npx))))
